fix: be more selective about div-wrapping
When HTML elements remain unclosed at the end of a raw HTML block, we
sometimes wrap the contents of the element in a <div> to preserve the
implicit structure when converting to e.g. EPUB for which Pandoc breaks
documents into sections based on inferred structure.

The old behavior was too aggressive and resulted in unintended line
breaks because pandoc has no notion of inline divs. This change refines
the heuristic to be more selective.