fix: correctly generate TOC when nested chapters don't contain H1s (#182)
Currently, mdbook-pandoc only generates table of contents (TOC) entries
for chapters that begin with an H1 (`# Chapter Heading`). Some books,
such as the rust book, use proportionally smaller headings in nested
chapters; for instance, chapter 1.1 begins with `## Installation`. In
order to correctly generate TOCs for books that follow this style, this
change adjusts heading levels based on the level of the first heading in
each chapter, so chapter 1.1 is treated as if it began with `#
Installation`.